requirements.txt
python-dotenv==1.1.0
openai==1.69.0
streamlit==1.44.0
